rename tool-wrapper to aisdk-tool-wrapper
