+++ packages/tools/src/providers/vercel-ai/quote.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { getFMPClient } from '@/types';
-import { createTool } from '@/utils/tool-wrapper';
+import { createTool } from '@/utils/aisdk-tool-wrapper';
 
 export const quoteTools = {
   getStockQuote: createTool({
